ai_pipeline/requirements.txt
pymupdf
google-generativeai
langchain-text-splitters
python-dotenv
tiktoken
